+++ src/scripts/create_psc.py
@@ -75,7 +75,7 @@ def main() -> None:
     )
 
     dataset_id = "EuroEval/psc-mini"
-    
+
     # Push the dataset to the Hugging Face Hub
     HfApi().delete_repo(dataset_id, repo_type="dataset", missing_ok=True)
     dataset.push_to_hub(dataset_id, private=True)